backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
channels
channels-redis
django-cors-headers
django-filter
djangorestframework-jwt
django-environ
asgiref